Refactor CollectionClient and PlaylistClient to import getQueueItemPlaybackStartTime from playerQueue
