+++ src/app/(main)/library/[library]/collection/[collection]/CollectionClient.tsx
@@ -17,7 +17,8 @@ import { useUser } from '@/contexts/UserContext'
 import { useCollectionBooks } from '@/hooks/useCollectionBooks'
 import { useCollectionDisplayMode } from '@/hooks/useCollectionDisplayMode'
 import { useTypeSafeTranslations } from '@/hooks/useTypeSafeTranslations'
-import { buildCollectionQueueItems, getPlayableCollectionBooks, getQueueItemPlaybackStartTime } from '@/lib/compilationPlayback'
+import { buildCollectionQueueItems, getPlayableCollectionBooks } from '@/lib/compilationPlayback'
+import { getQueueItemPlaybackStartTime } from '@/lib/playerQueue'
 import { Collection } from '@/types/api'
 import { useRouter } from 'next/navigation'
 import { useCallback, useEffect, useMemo, useState } from 'react'
