+++ src/app/(main)/library/[library]/playlist/[playlist]/PlaylistClient.tsx
@@ -16,7 +16,8 @@ import { useUser } from '@/contexts/UserContext'
 import { usePlaylistDisplayMode } from '@/hooks/usePlaylistDisplayMode'
 import { usePlaylistItems } from '@/hooks/usePlaylistItems'
 import { useTypeSafeTranslations } from '@/hooks/useTypeSafeTranslations'
-import { buildPlaylistQueueItems, getPlayablePlaylistItems, getQueueItemPlaybackStartTime } from '@/lib/compilationPlayback'
+import { buildPlaylistQueueItems, getPlayablePlaylistItems } from '@/lib/compilationPlayback'
+import { getQueueItemPlaybackStartTime } from '@/lib/playerQueue'
 import type { Playlist } from '@/types/api'
 import { useRouter } from 'next/navigation'
 import { useCallback, useEffect, useMemo, useState } from 'react'
